修改:上报接口 和 dialog弹框

+++ app/src/main/java/com/rz/smart/ui/main/MainFragment.kt
@@ -94,7 +94,7 @@ class MainFragment : Fragment() {
                                 entity.GoodsID,
                                 entity.SupplierID?.toInt()!!,
                                 entity.GoodsWeight,
-                                goodsAmount.toInt(),
+                                goodsAmount,
                                 entit2y.WarehouseID,
                                 "${CacheDataUtils.USERNAME1!!},${CacheDataUtils.USERNAME2!!}"
                             )
